🎨 Consolidate errors into one file
- This should leave errors largely unchanged
- names should be the same
- I've dropped some custom error messages to streamline things
- AFAIK these types are not used outside of this package so this shouldn't have much effect
- Why? I'm trying to move more code from ./tdf3 to ./src
