--- lib/src/nanotdf/errors/InvalidEphemeralKeyError.ts
@@ -1,12 +1,0 @@
-class InvalidEphemeralKeyError extends Error {
-  __proto__: Error;
-
-  constructor() {
-    const trueProto = new.target.prototype;
-    super('Incorrect length for ephemeral key');
-    this.__proto__ = trueProto;
-    this.name = 'InvalidEphemeralKeyError';
-  }
-}
-
-export default InvalidEphemeralKeyError;
