+++ lib/tdf3/src/client/validation.ts
@@ -1,4 +1,4 @@
-import { AttributeValidationError, IllegalArgumentError } from '../errors.js';
+import { AttributeValidationError, IllegalArgumentError } from '../../../src/errors.js';
 import { AttributeObject } from '../models/attribute-set.js';
 
 const sageGetMatch = (match: RegExpMatchArray | null) => (match ? match[0] : null);
