+++ lib/tdf3/src/index.ts
@@ -1,5 +1,5 @@
 export * as Client from './client/index.js';
 export { Client as TDF3Client } from './client/index.js';
-export * as Errors from './errors.js';
+export * as Errors from '../../src/errors.js';
 export { TDF } from './tdf.js';
 export { version, clientType } from './version.js';
