+++ lib/tests/mocha/unit/errors.spec.ts
@@ -9,7 +9,7 @@ import {
   TdfDecryptError,
   TdfError,
   TdfPayloadExtractionError,
-} from '../../../tdf3/src/errors.js';
+} from '../../../src/errors.js';
 
 describe('Errors', () => {
   const errorClasses: Record<string, typeof TdfError> = {
